refactor : 초기생성 완성본
