+++ src/main/java/com/clean/cleanroom/estimate/entity/Estimate.java
@@ -1,12 +1,11 @@
 package com.clean.cleanroom.estimate.entity;
 
 import com.clean.cleanroom.commission.entity.Commission;
-import com.clean.cleanroom.pirtner.entity.Partner;
+import com.clean.cleanroom.partner.entity.Partner;
 import jakarta.persistence.*;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 @Getter
 @Entity
@@ -27,7 +26,7 @@ public class Estimate {
     private int price;
 
     @Column(nullable = true)
-    private LocalDateTime fixed_date;
+    private LocalDateTime fixedDate;
 
     @Column(nullable = true, length = 1000)
     private String statement;
