+++ src/main/java/com/clean/cleanroom/partner/controller/PartnerController.java
@@ -1,0 +1,4 @@
+package com.clean.cleanroom.partner.controller;
+
+public class PartnerController {
+}
