+++ src/main/java/com/clean/cleanroom/partner/dto/PartnerResponseDto.java
@@ -1,0 +1,4 @@
+package com.clean.cleanroom.partner.dto;
+
+public class PartnerResponseDto {
+}
