+++ src/main/java/com/clean/cleanroom/partner/entity/Partner.java
@@ -1,0 +1,17 @@
+package com.clean.cleanroom.partner.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+@Entity
+public class Partner {
+
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String phoneNumber;
+}
