+++ src/main/java/com/clean/cleanroom/partner/repository/PartnerRepository.java
@@ -1,0 +1,7 @@
+package com.clean.cleanroom.partner.repository;
+
+import com.clean.cleanroom.partner.entity.Partner;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PartnerRepository extends JpaRepository<Partner, Long> {
+}
