+++ src/main/java/com/clean/cleanroom/partner/service/PartnerService.java
@@ -1,0 +1,4 @@
+package com.clean.cleanroom.partner.service;
+
+public class PartnerService {
+}
