--- src/main/java/com/clean/cleanroom/pirtner/controller/PartnerController.java
@@ -1,4 +1,0 @@
-package com.clean.cleanroom.pirtner.controller;
-
-public class PartnerController {
-}
