--- src/main/java/com/clean/cleanroom/pirtner/dto/PartnerRequestDto.java
@@ -1,4 +1,0 @@
-package com.clean.cleanroom.pirtner.dto;
-
-public class PartnerRequestDto {
-}
