--- src/main/java/com/clean/cleanroom/pirtner/dto/PartnerResponseDto.java
@@ -1,4 +1,0 @@
-package com.clean.cleanroom.pirtner.dto;
-
-public class PartnerResponseDto {
-}
