--- src/main/java/com/clean/cleanroom/pirtner/entity/Partner.java
@@ -1,17 +1,0 @@
-package com.clean.cleanroom.pirtner.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class Partner {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String phoneNumber;
-}
