--- src/main/java/com/clean/cleanroom/pirtner/service/PartnerService.java
@@ -1,4 +1,0 @@
-package com.clean.cleanroom.pirtner.service;
-
-public class PartnerService {
-}
